Use lookup table
git-svn-id: svn+ssh://solsvn.oma.be/home/sidc/ssr/SWHV_repos/trunk/client/jhv-3d-wcs@3081 b4e469a2-07ce-4b26-9273-4d7d95a670c7

+++ src/plugins/jhv-eveplugin/src/org/helioviewer/plugins/eveplugin/radio/data/DownloadedJPXDataWorkerResult.java
@@ -3,12 +3,12 @@ package org.helioviewer.plugins.eveplugin.radio.data;
 import java.awt.Rectangle;
 
 public class DownloadedJPXDataWorkerResult {
-    private final byte[] data;
+    private final int[] data;
     private final long imageID;
     private final long downloadID;
     private final Rectangle dataSize;
 
-    public DownloadedJPXDataWorkerResult(byte[] data, long imageID, long downloadID, Rectangle dataSize) {
+    public DownloadedJPXDataWorkerResult(int[] data, long imageID, long downloadID, Rectangle dataSize) {
         super();
         this.data = data;
         this.imageID = imageID;
@@ -19,7 +19,7 @@ public class DownloadedJPXDataWorkerResult {
     /**
      * @return the data
      */
-    public byte[] getData() {
+    public int[] getData() {
         return data;
     }
